- Update UISate code - Update compose version 1.8.1 to 1.8.2 - Update lifecycle version 2.8.0 to 2.8.3
