+++ composeApp/src/commonMain/kotlin/ui/celebrities/CelebrityUiState.kt
@@ -1,0 +1,9 @@
+package ui.celebrities
+
+import data.model.celebrities.Celebrity
+
+data class CelebrityUiState(
+    val celebrityList: List<Celebrity>? = emptyList(),
+    val isLoading: Boolean = false,
+    val errorMessage: String? = null
+)
