+++ composeApp/src/commonMain/kotlin/ui/movie/MovieUiState.kt
@@ -1,0 +1,9 @@
+package ui.movie
+
+import data.model.MovieItem
+
+data class MovieUiState(
+    val movieList: List<MovieItem>? = emptyList(),
+    val isLoading: Boolean = false,
+    val errorMessage: String? = null
+)
