+++ composeApp/src/commonMain/kotlin/ui/movie/detail/MovieDetailViewModel.kt
@@ -8,7 +8,6 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
-import utils.MovieDetailUiState
 import utils.network.UiState
 
 class MovieDetailViewModel(private val repo: Repository = Repository()) : ViewModel() {
