+++ composeApp/src/commonMain/kotlin/ui/movie/top_rated/TopRatedViewModel.kt
@@ -8,7 +8,7 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
-import utils.MovieUiState
+import ui.movie.MovieUiState
 import utils.Paginator
 
 class TopRatedViewModel(private val repo: Repository = Repository()) : ViewModel() {
