+++ composeApp/src/commonMain/kotlin/ui/tv_series/TvSeriesUiState.kt
@@ -1,0 +1,9 @@
+package ui.tv_series
+
+import data.model.TvSeriesItem
+
+data class TvSeriesUiState(
+    val tvSeriesList: List<TvSeriesItem>? = emptyList(),
+    val isLoading: Boolean = false,
+    val errorMessage: String? = null
+)
