+++ composeApp/src/commonMain/kotlin/ui/tv_series/detail/TvSeriesDetailViewModel.kt
@@ -8,7 +8,6 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
-import utils.TvSeriesDetailUiState
 import utils.network.UiState
 
 class TvSeriesDetailViewModel(private val repo: Repository = Repository()) : ViewModel() {
